requirements.txt
aiogram==3.25.0
aiohttp==3.11.0
sqlalchemy==2.0.23
Pillow==10.1.0
python-dotenv==1.0.0